created file to run multiple scripts
there seem to be issues with running multiple scripts at the same time (probably cause of emu.frameadvance()) so i compile my scripts into one where they can be easily turned on and off. There's a bunch of crap in here of course but I haven't noticed any lag yet. Just change any of the booleans at the top to "false" to disable that script